Add files via upload
As number of enthusiastic students increase drastically, day-by-day, lecturers are now being burdened with the load of, not only teaching the huge masses, but also correcting their quiz papers. Hence to reduce this problem “ONLINE TEST” has been designed us. Here a student has to login and answer the questions pre-fed by lecturers, in the computer. Results are calculated instantaneously as and when the student answers the questions and shown at the end of session.

The web application we have developed makes use of mysql  to manage the databases contaning login details, correct answers will be stored in “Locked Tables” hence making the application secure.
HTML along with PHP is used and hence making our system reliable.

Concluding, the application not only easens the work load on the lectures but is also user friendly and reduces  the usage of papers, other utilities and making the application an efficient one.
.